backend/pyproject.toml
[project]
name = "kyudai-campus-sns-backend"
version = "0.1.0"
description = "Kyudai Campus SNS Backend API"
authors = [
    { name = "AI Developer", email = "ai@devcontainer" }
]
dependencies = [
    "fastapi==0.109.0",
    "uvicorn[standard]==0.27.0",
    "asyncpg==0.29.0",
    "sqlalchemy==2.0.25",
    "pydantic==2.5.3",
    "python-multipart==0.0.6",
    "python-ulid==2.2.0",
]
requires-python = ">=3.11"

[project.optional-dependencies]
dev = [
    "pytest==7.4.4",
    "pytest-asyncio==0.23.3",
    "httpx==0.26.0",
    "black",
    "ruff",
    "ipython",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.black]
line-length = 100
target-version = ['py311']

[tool.ruff]
line-length = 100
select = ["E", "F", "W", "I", "N"]
ignore = ["E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
addopts = "-v"
asyncio_mode = "auto"